Add support for local configuration files
It is now possible to create a config.local.php file that will be used
to pull in settings rather than modifying the config.default.php. This
local file is ignored by git.

To create a local file copy the default file into a new one.

    $ cp config.default.php config.local.php

Then edit the settings at will.