Auto increment now works in Firefox and Opera. Fixes #182
Firefox was not sending the _csrf cookie when the save request was
made unless the version number was included in the url string. Not
100% sure why it's a very odd behaviour but this caused the request to
fail due to an invalid CSRF validation.

The solution was to always include the revision number in the url,
and the update was made to the $code_id in the PHP app. However this
issue was introduced with changes in commit 601caf48 and may have
broken other functionality.