Stupid fix for Firefox
Basically because the element vanishes from under the mouse, the item
the user thought they were clicking on is no longer there, so in the
end, no link is clicked, so nothing happens. This change just defers
the menu from being hidden on the next tick - most importantly: /after/
the click has happened.