I think this solves the entire problem if IE8-9 not having history API support
If there's no history API, just update the location.hash, but all the
other UI components update as if the real url were used. If the user
copies and pastes the url to twitter (say for instance), the
hashbang-jump (though really there's no bang) detects the *real* url on
the hash, and simple redirects. This is made the first part of the
compressed jsbin…min.js so it should happen pretty early.