Add Zen Coding to HTML pane, addresses #285.
Using zen-coding for CodeMirror I've added Zen Coding support to
the HTML pane of JSBin.

It's on by default, with no way to switch it off. I'm not sure how
it's best to do that – from looking at the TODOs, are CodeMirror
options something that's planned?

In Opera, becuase 'tab' is the Zen Coding trigger, the line that
the user ends up on is also indented. This is not the case in
Firefox and Chrome.