Menu help shows in navigation
So this is an experiment based on issue #729 to see if I can make it clear how each feature works. Each menu item had a tooltip, but I personally never found them useful because the mouse never hovers long enough and then it automatically goes away. Maybe this is a good change?

My only concern is it might annoy the pro users, so maybe there's a way of turning it off?