Stop duplicate copies of React from auto-adding
Previously, selecting "JSX (React)" could cause react-0.8.0.js to be added, even if the add-ons build or a different version was already included by hand. Now we won't add it if the HTML contains anything that looks like a React script tag.

Also changed it to use unminified builds for better warnings and to default to the add-ons build because it's more general.

Test Plan:
In a new bin, selected "JSX (React)" and saw that the add-ons script tag was added. In a separate bin, added a script tag for react-0.9.0-rc1.js, then selected JSX and the HTML didn't change.