Limit full previews for 2 hours
We've been getting more and more spammers using JS Bin to create either phishing pages, or using the JS to redirect. We have an internal method to blacklist these spammers, but JS Bin's weak spot is that anyone can create anonymous bins and link to the full output (i.e. without any of the JS Bin bits - like a mini geocities).

It's been getting worse as time goes on, so I'm moving to a model that allows full previews for 2 hours, then automatically redirects to the /edit view - which at the very least, informs the visitor that they're actually on jsbin.com and not being phished.

I will adjust this change so that it only applies to *new* bins since the release of this code.

If you create your bins as registered user, this has no effect for you.

However, if you're registered and creating spam, it'll in breach of our T&Cs, so we will remove your account - be nice!