Fix #1633
First test for the meta tag with a charset property, and insert the license copy below.

If that's not found, put it after the html element. Sure, we should probably look for the title tag, but let's face it, it's just a dumping ground for test code, and dear author: include a meta charset for heavens' sake.

...okay, maybe I should do the title check too, but maybe _after_ someone files an issue.