Method to detect previously crashed state.
When the browser renders the output with live JS enabled, it sets a flag, at's then removed (directly after we finish writing the contents of the iframe). *Or* the can be removed by the user changing their code.

If the page is loaded, and the flag is present, that means the page somehow crashed during load time, so we automatically turn off live-js for the user and show them a message that we've done so.