Add <base> to rendered frame.
This is an important change, because we're seeing /a lot/ of bins being created that contain a lot of images, and when they're using relative urls (or just placeholder text like "xxx") the images request jsbin.com, _with each keystroke_.

The knock on effect is that XHR requests to jsbin (for joining jsbins together) will fail if they're using relative urls.

I'd like to get feedback on the impact of this change, because it'll alleviate a lot of unnecessary pressure from our servers.