First pass at aria-label
I ran this with both textarea mode *and* CodeMirror mode and the label remains in place after CodeMirror has finished loading.

I'm not 100% sure how the screen reader is going to use the aria label for this textarea (equally, I'm not sure how a screen reader is going to use CodeMirror in the first place).

Input welcome.