fix: support posting again
Since we moved to a separate loader for the bin content, it meant that the 2nd http request (for start.js) didn't know anything about the POSTed values sent to the editor page. I tried to store it originally on the `req.flash` but the example pages from the FT were great because they were larger than the cookie could handle (and thus were chomped and discarded).

It took me a looooong time to work it out, but actually a simple memory store fixes the issue. Similar to a flash that's stored on the cookie, but this is a time sensitive flash that's kept in memory, and removed as soon as it's retrieved (but also removed after a 5 second timeout). Since it's an edge case of JS Bin, I'm happy that it's in memory (though I will have to keep an eye on it too).