Add jQuery & jQuery Compat 3.0.0-alpha1
I'll be submitting PRs changing this to further 3.0.0 alphas/betas so that people can test the latest released preview of the upcoming jQueries 3.0.0.

BTW, I've noticed that some of the links direct to minified files and some to full ones (like jQuery UI). Why is that? Wouldn't it be better if all links were non-minified? It's just for testing anyway.