Upgrade to latest sqlite3 version
This updates the sqlite3 dependency to the latest available version.
Among other things, this should enable people interested in hosting
their own copy of jsbin to specify binary mirrors of the sqlite3
dependency. The reason this requires an update is that the current
version being used depends on a version of node-pre-gyp that doesn't
support mirroring – see mapbox/node-pre-gyp#170 for more info on
this feature.