fix: embeds rendering again
Fixes #3047

For some reason, the live.js renderLivePreview function doesn't have
access to the "real" (and full) jsbin object, so it was missing the user
which I added an arb check for. It wasn't needed, so I reverted the
change but it's still odd that it's not available at that time.